+++ packages/lumx-angularjs/src/components/input-label/input-label_directive.js
@@ -64,10 +64,10 @@ function InputLabelDirective() {
         replace: true,
         restrict: 'E',
         scope: {
-            label: '@lxLabel',
             theme: '@?lxTheme',
         },
         template,
+        transclude: true,
     };
 }