+++ packages/lumx-angularjs/src/components/switch/switch_directive.js
@@ -151,33 +151,16 @@ function SwitchDirective() {
 
         if (transclude.isSlotFilled('label')) {
             ctrls[0].hasLabel = true;
-
-            transclude(
-                (clone) => {
-                    ctrls[0].label = clone.text().trim();
-                },
-                null,
-                'label',
-            );
         }
 
         if (transclude.isSlotFilled('helper')) {
             ctrls[0].hasHelper = true;
-
-            transclude(
-                (clone) => {
-                    ctrls[0].helper = clone.text().trim();
-                },
-                null,
-                'helper',
-            );
         }
 
         if (!ctrls[0].hasLabel && !ctrls[0].hasHelper) {
             transclude((clone) => {
                 if (clone.length > 0) {
                     ctrls[0].hasTranscluded = true;
-                    ctrls[0].label = clone.text().trim();
                 }
             });
         }